docs: reestructurar documentación y modularizar transcriber
- Renombrar documentacion/ → docs/ (convención estándar)
- Mover MEJORAS.md → docs/mejoras.md
- Crear CHANGELOG.md con historial versionado (Keep a Changelog)
- Actualizar links en README.md
- Extraer transcriptor de audio a server/transcriber.js desde telegram.js
